backend/requirements.txt
flask
flask-socketio
digi-xbee
eventlet
